Implemented everything for regex inversion, added tests

+++ src/engine/automaton-graphviz-printer.cpp
@@ -9,14 +9,14 @@ std::ostream &operator<<(std::ostream &stream, const AutomatonGraphvizPrinter &p
 
     auto &states = printer.automaton.get_states();
 
-    for (int i = 0; i < states.size(); i++) {
+    for (size_t i = 0; i < states.size(); i++) {
         if (states[i].is_final) {
             stream << i << " ";
         }
     }
     stream << ";" << std::endl;
     stream << "    node [shape = circle];" << std::endl;
-    for (int i = 0; i < states.size(); i++) {
+    for (size_t i = 0; i < states.size(); i++) {
         for (auto &transition: states[i].transitions) {
             stream << "    " << i << " -> " << transition.target_index << " [ label = \"" << transition.regex
                    << "\" ];" << std::endl;
